+++ 301-vm-32-data-disks-high-iops/DSC/StoragePool.ps1
@@ -1,0 +1,18 @@
+Configuration StoragePool
+{
+  param ($MachineName)
+
+  Node $MachineName
+  {
+	Script ConfigureStoragePool { 
+		SetScript = { 
+			$disks = Get-PhysicalDisk –CanPool $true
+			New-StoragePool -FriendlyName "DataPool" -StorageSubsystemFriendlyName "Windows Storage*" -PhysicalDisks $disks | New-VirtualDisk -FriendlyName "DataDisk" -UseMaximumSize -NumberOfColumns $disks.Count -ResiliencySettingName "Simple" -ProvisioningType Fixed -Interleave 65536 | Initialize-Disk -Confirm:$False -PassThru | New-Partition -DriveLetter H –UseMaximumSize | Format-Volume -FileSystem NTFS -NewFileSystemLabel "DATA" -Confirm:$false			
+		} 
+
+		TestScript = { 
+			Test-Path H:\ 
+		} 
+		GetScript = { <# This must return a hash table #> }          }   
+  }
+}