+++ bosh-setup/scripts/setup_env.sh
@@ -2,65 +2,60 @@
 
 source utils.sh
 
-echo "Start to update package lists from repositories..."
-retryop "apt-get update"
-
-echo "Start to install prerequisites..." 
-retryop "apt-get -y install build-essential zlibc zlib1g-dev ruby ruby-dev openssl libxslt-dev libxml2-dev libssl-dev libreadline6 libreadline6-dev libyaml-dev libsqlite3-dev sqlite3 python-dev python-pip jq"
-
 set -e
 
-tenant_id=$1
-client_id=$2
-base64_encoded_client_secret_or_certificate=$3
+echo "Installing jq"
+retryop "apt-get update && apt-get install -y jq"
+
 custom_data_file="/var/lib/cloud/instance/user-data.txt"
 settings=$(cat ${custom_data_file})
-
 function get_setting() {
   key=$1
   local value=$(echo $settings | jq ".$key" -r)
   echo $value
 }
 
-function install_bosh_cli() {
-  echo "Start to install bosh-cli v2..."
-  bosh_cli_url=$1
-  wget $bosh_cli_url
-  chmod +x ./bosh-cli-*
-  sudo mv ./bosh-cli-* /usr/local/bin/bosh
-}
-
+# Service Principal
+environment=$(get_setting ENVIRONMENT)
+service_principal_type=$(get_setting SERVICE_PRINCIPAL_TYPE)
+tenant_id=$1
+client_id=$2
+base64_encoded_client_secret_or_certificate=$3
 function client_secret_or_certificate() {
   echo ${base64_encoded_client_secret_or_certificate} | base64 --decode
 }
 
-environment=$(get_setting ENVIRONMENT)
-service_principal_type=$(get_setting SERVICE_PRINCIPAL_TYPE)
+# https://bosh.io/docs/cli-v2-install/#additional-dependencies
+echo "Installing OS specified dependencies for bosh create-env command"
+retryop "apt-get update && apt-get install -y build-essential zlibc zlib1g-dev ruby ruby-dev openssl libxslt-dev libxml2-dev libssl-dev libreadline6 libreadline6-dev libyaml-dev libsqlite3-dev sqlite3"
 
-set +e
+echo "Installing BOSH CLI"
+bosh_cli_url=$(get_setting BOSH_CLI_URL)
+wget $bosh_cli_url
+chmod +x ./bosh-cli-*
+mv ./bosh-cli-* /usr/local/bin/bosh
 
-echo "Start to install python packages..."
-pkg_list="setuptools==32.3.1 azure==2.0.0rc1"
-if [ "$environment" = "AzureChinaCloud" ]; then
-  for pkg in $pkg_list; do
-    retryop "pip install $pkg --index-url https://mirror.azure.cn/pypi/simple/ --default-timeout=60"
-  done
-else
-  for pkg in $pkg_list; do
-    retryop "pip install $pkg"
-  done
-fi
-
-set -e
+# https://docs.microsoft.com/en-us/cli/azure/install-azure-cli-apt?view=azure-cli-latest#install
+echo "Installing Azure CLI"
+AZ_REPO=$(lsb_release -cs)
+echo "deb [arch=amd64] https://packages.microsoft.com/repos/azure-cli/ $AZ_REPO main" | tee /etc/apt/sources.list.d/azure-cli.list
+curl -L https://packages.microsoft.com/keys/microsoft.asc | sudo apt-key add -
+retryop "apt-get install apt-transport-https"
+retryop "apt-get update && apt-get install azure-cli=2.0.33-1~$AZ_REPO"
 
 echo "Creating the containers (bosh and stemcell) and the table (stemcells) in the default storage account"
 default_storage_account=$(get_setting DEFAULT_STORAGE_ACCOUNT_NAME)
 default_storage_access_key=$(get_setting DEFAULT_STORAGE_ACCESS_KEY)
 endpoint_suffix=$(get_setting SERVICE_HOST_BASE)
-python prepare_storage_account.py ${default_storage_account} ${default_storage_access_key} ${endpoint_suffix} ${environment}
-
-bosh_cli_url=$(get_setting BOSH_CLI_URL)
-install_bosh_cli $bosh_cli_url
+connection_string="DefaultEndpointsProtocol=https;AccountName=${default_storage_account};AccountKey=${default_storage_access_key};EndpointSuffix=${endpoint_suffix}"
+if [ "$environment" = "AzureStack" ]; then
+  cat /var/lib/waagent/Certificates.pem >> /etc/ssl/certs/ca-certificates.crt
+  export REQUESTS_CA_BUNDLE=/etc/ssl/certs/ca-certificates.crt
+  az cloud update --profile 2017-03-09-profile
+fi
+az storage container create --name bosh --connection-string ${connection_string}
+az storage container create --name stemcell --connection-string ${connection_string}
+az storage table create --name stemcells --connection-string ${connection_string}
 
 username=$(get_setting ADMIN_USER_NAME)
 home_dir="/home/$username"
@@ -117,18 +112,21 @@ cat > "$home_dir/deploy_bosh.sh" << EOF
 
 set -e
 
-export BOSH_LOG_LEVEL="debug"
+export BOSH_LOG_LEVEL="$(get_setting LOG_LEVEL_FOR_BOSH)"
 export BOSH_LOG_PATH="./run.log"
 
 bosh create-env ~/example_manifests/bosh.yml \\
   --state=state.json \\
   --vars-store=~/bosh-deployment-vars.yml \\
   -o ~/example_manifests/cpi.yml \\
+  -o ~/example_manifests/use-location.yml \\
   -o ~/example_manifests/custom-cpi-release.yml \\
   -o ~/example_manifests/custom-environment.yml \\
   -o ~/example_manifests/use-azure-dns.yml \\
   -o ~/example_manifests/jumpbox-user.yml \\
   -o ~/example_manifests/keep-failed-or-unreachable-vms.yml \\
+  -o ~/example_manifests/uaa.yml \\
+  -o ~/example_manifests/credhub.yml \\
   -v director_name=azure \\
   -v internal_cidr=10.0.0.0/24 \\
   -v internal_gw=10.0.0.1 \\
@@ -137,6 +135,7 @@ bosh create-env ~/example_manifests/bosh.yml \\
   -v cpi_release_sha1=$(get_setting BOSH_AZURE_CPI_RELEASE_SHA1) \\
   -v director_vm_instance_type=$(get_setting BOSH_VM_SIZE) \\
   -v resource_group_name=$(get_setting RESOURCE_GROUP_NAME) \\
+  -v location=$(get_setting LOCATION) \\
   -v vnet_name=$(get_setting VNET_NAME) \\
   -v subnet_name=$(get_setting SUBNET_NAME_FOR_BOSH) \\
   -v default_security_group=$(get_setting NSG_NAME_FOR_BOSH) \\
@@ -154,6 +153,12 @@ elif [ "${service_principal_type}" == "Certificate" ]; then
   cat >> "$home_dir/deploy_bosh.sh" << EOF
   -o ~/example_manifests/use-service-principal-with-certificate.yml \\
   -l ~/example_manifests/service-principal-certificate.yml \\
+EOF
+fi
+
+if [ "$environment" = "AzureStack" ]; then
+  cat >> "$home_dir/deploy_bosh.sh" << EOF
+  -o ~/example_manifests/use-trusted-certs.yml \\
 EOF
 fi
 
@@ -177,9 +182,50 @@ else
   -o ~/example_manifests/use-managed-disks.yml
 EOF
 fi
+
+cat >> "$home_dir/deploy_bosh.sh" << EOF
+
+cat >> "$home_dir/.profile" << EndOfFile
+# BOSH CLI
+export BOSH_ENVIRONMENT=10.0.0.4
+export BOSH_CLIENT=admin
+export BOSH_CLIENT_SECRET="\$(bosh int ~/bosh-deployment-vars.yml --path /admin_password)"
+export BOSH_CA_CERT="\$(bosh int ~/bosh-deployment-vars.yml --path /director_ssl/ca)"
+EndOfFile
+source $home_dir/.profile
+EOF
+
 chmod 777 $home_dir/deploy_bosh.sh
 
+cat > "$home_dir/login_bosh.sh" << EOF
+#!/usr/bin/env bash
+
+export BOSH_ENVIRONMENT=10.0.0.4
+export BOSH_CLIENT=admin
+export BOSH_CLIENT_SECRET="\$(bosh int ~/bosh-deployment-vars.yml --path /admin_password)"
+export BOSH_CA_CERT="\$(bosh int ~/bosh-deployment-vars.yml --path /director_ssl/ca)"
+
+bosh alias-env azure
+bosh -e azure login
+EOF
+chmod 777 $home_dir/login_bosh.sh
+
+cat > "$home_dir/connect_director_vm.sh" << EOF
+#!/usr/bin/env bash
+
+bosh int ~/bosh-deployment-vars.yml --path /jumpbox_ssh/private_key > jumpbox.key
+chmod 600 jumpbox.key
+ssh jumpbox@10.0.0.4 -i jumpbox.key
+EOF
+chmod 777 $home_dir/connect_director_vm.sh
+
+system_domain=$(get_setting SYSTEM_DOMAIN)
+if [ "${system_domain}" = "NotConfigured" ]; then
+  system_domain="$(get_setting CLOUD_FOUNDRY_PUBLIC_IP).xip.io"
+fi
 cat > "$home_dir/deploy_cloud_foundry.sh" << EOF
+#!/usr/bin/env bash
+
 export BOSH_ENVIRONMENT=10.0.0.4
 export BOSH_CLIENT=admin
 export BOSH_CLIENT_SECRET="\$(bosh int ~/bosh-deployment-vars.yml --path /admin_password)"
@@ -195,78 +241,68 @@ bosh -n update-cloud-config ~/example_manifests/cloud-config.yml \\
   -v subnet_name=$(get_setting SUBNET_NAME_FOR_CLOUD_FOUNDRY) \\
   -v security_group=$(get_setting NSG_NAME_FOR_CLOUD_FOUNDRY) \\
   -v load_balancer_name=$(get_setting LOAD_BALANCER_NAME)
-
-bosh upload-stemcell --sha1=$(get_setting STEMCELL_SHA1) $(get_setting STEMCELL_URL)
 EOF
 
+if [ "$environment" = "AzureChinaCloud" ]; then
+  cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
+bosh -n update-runtime-config ~/example_manifests/dns.yml \\
+  -o ~/example_manifests/use-mirror-bosh-dns-release.yml \\
+  --name=dns
+EOF
+else
+  cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
+bosh -n update-runtime-config ~/example_manifests/dns.yml \\
+  --name=dns
+EOF
+fi
+
 cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
+bosh upload-stemcell --sha1=$(get_setting STEMCELL_SHA1) $(get_setting STEMCELL_URL)
+
 bosh -n -d cf deploy ~/example_manifests/cf-deployment.yml \\
   --vars-store=~/cf-deployment-vars.yml \\
   -o ~/example_manifests/azure.yml \\
   -o ~/example_manifests/scale-to-one-az.yml \\
 EOF
+
 if [ "$environment" = "AzureChinaCloud" ]; then
   cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
-  -o ~/example_manifests/use-azure-storage-blobstore.yml \\
-  -o ~/example_manifests/use-mirror-releases-for-cf.yml \\
-  -v system_domain=$(get_setting CLOUD_FOUNDRY_PUBLIC_IP).xip.io \\
-  -v environment=$(get_setting ENVIRONMENT) \\
-  -v blobstore_storage_account_name=$(get_setting DEFAULT_STORAGE_ACCOUNT_NAME) \\
-  -v blobstore_storage_access_key=$(get_setting DEFAULT_STORAGE_ACCESS_KEY) \\
-  -v app_package_directory_key=cc-packages \\
-  -v buildpack_directory_key=cc-buildpack \\
-  -v droplet_directory_key=cc-droplet \\
-  -v resource_directory_key=cc-resource
-EOF
-elif [ "$environment" = "AzureStack" ]; then
-  cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
-  -o ~/example_manifests/use-compiled-releases.yml \\
-  -v system_domain=$(get_setting CLOUD_FOUNDRY_PUBLIC_IP).xip.io
+  -o ~/example_manifests/use-mirror-compiled-releases.yml \\
 EOF
 else
   cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
   -o ~/example_manifests/use-compiled-releases.yml \\
+EOF
+fi
+
+cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
+  -o ~/example_manifests/use-external-blobstore.yml \\
+  -v app_package_directory_key=cc-packages \\
+  -v buildpack_directory_key=cc-buildpacks \\
+  -v droplet_directory_key=cc-droplets \\
+  -v resource_directory_key=cc-resources \\
   -o ~/example_manifests/use-azure-storage-blobstore.yml \\
-  -v system_domain=$(get_setting CLOUD_FOUNDRY_PUBLIC_IP).xip.io \\
   -v environment=$(get_setting ENVIRONMENT) \\
   -v blobstore_storage_account_name=$(get_setting DEFAULT_STORAGE_ACCOUNT_NAME) \\
   -v blobstore_storage_access_key=$(get_setting DEFAULT_STORAGE_ACCESS_KEY) \\
-  -v app_package_directory_key=cc-packages \\
-  -v buildpack_directory_key=cc-buildpack \\
-  -v droplet_directory_key=cc-droplet \\
-  -v resource_directory_key=cc-resource
 EOF
-fi 
+if [ "$environment" = "AzureStack" ]; then
+  cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
+  -o ~/example_manifests/use-azure-stack-storage-blobstore.yml \\
+  -v blobstore_storage_dns_suffix=${endpoint_suffix} \\
+EOF
+fi
+cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
+  -v system_domain=${system_domain}
+EOF
 chmod 777 $home_dir/deploy_cloud_foundry.sh
 
-cat >> "$home_dir/connect_director_vm.sh" << EOF
-#!/usr/bin/env bash
-
-bosh int ~/bosh-deployment-vars.yml --path /jumpbox_ssh/private_key > jumpbox.key
-chmod 600 jumpbox.key
-ssh jumpbox@10.0.0.4 -i jumpbox.key
-EOF
-chmod 777 $home_dir/connect_director_vm.sh
-
-cat >> "$home_dir/login_bosh.sh" << EOF
-#!/usr/bin/env bash
-
-export BOSH_ENVIRONMENT=10.0.0.4
-export BOSH_CLIENT=admin
-export BOSH_CLIENT_SECRET="\$(bosh int ~/bosh-deployment-vars.yml --path /admin_password)"
-export BOSH_CA_CERT="\$(bosh int ~/bosh-deployment-vars.yml --path /director_ssl/ca)"
-
-bosh alias-env azure
-bosh -e azure login
-EOF
-chmod 777 $home_dir/login_bosh.sh
-
-cat >> "$home_dir/login_cloud_foundry.sh" << EOF
+cat > "$home_dir/login_cloud_foundry.sh" << EOF
 #!/usr/bin/env bash
 
 cf_admin_password="\$(bosh int ~/cf-deployment-vars.yml --path /cf_admin_password)"
 
-cf login -a https://api.$(get_setting CLOUD_FOUNDRY_PUBLIC_IP).xip.io -u admin -p "\${cf_admin_password}" --skip-ssl-validation
+cf login -a https://api.${system_domain} -u admin -p "\${cf_admin_password}" --skip-ssl-validation
 EOF
 chmod 777 $home_dir/login_cloud_foundry.sh