+++ mongodb-high-availability/mongodb-ubuntu-install.sh
@@ -31,6 +31,7 @@
 
 PACKAGE_URL=http://repo.mongodb.org/apt/ubuntu
 PACKAGE_NAME=mongodb-org
+PACKAGE_VERSION="4.0.2"
 REPLICA_SET_KEY_DATA=""
 REPLICA_SET_NAME=""
 REPLICA_SET_KEY_FILE="/etc/mongo-replicaset-key"
@@ -53,6 +54,7 @@ help()
 	echo "Options:"
 	echo "		-i Installation package URL"
 	echo "		-b Installation package name"
+	echo "		-v Installation package version"
 	echo "		-r Replica set name"
 	echo "		-k Replica set key"
 	echo "		-u System administrator's user name"
@@ -80,7 +82,7 @@ then
 fi
 
 # Parse script parameters
-while getopts :i:b:r:k:u:p:x:n:alh optname; do
+while getopts :i:b:v:r:k:u:p:x:n:alh optname; do
 
 	# Log input parameters (except the admin password) to facilitate troubleshooting
 	if [ ! "$optname" == "p" ] && [ ! "$optname" == "k" ]; then
@@ -93,6 +95,9 @@ while getopts :i:b:r:k:u:p:x:n:alh optname; do
 		;;
 	b) # Installation package name
 		PACKAGE_NAME=${OPTARG}
+		;;
+	v) # Installation package version
+		PACKAGE_VERSION=${OPTARG}
 		;;
 	r) # Replica set name
 		REPLICA_SET_NAME=${OPTARG}
@@ -114,7 +119,6 @@ while getopts :i:b:r:k:u:p:x:n:alh optname; do
 		;;		
 	a) # Arbiter indicator
 		IS_ARBITER=true
-		JOURNAL_ENABLED=false
 		;;		
 	l) # Last member indicator
 		IS_LAST_MEMBER=true
@@ -177,8 +181,8 @@ install_mongodb()
 	log "Downloading MongoDB package $PACKAGE_NAME from $PACKAGE_URL"
 
 	# Configure mongodb.list file with the correct location
-	apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 7F0CEB10
-	echo "deb ${PACKAGE_URL} "$(lsb_release -sc)"/mongodb-org/3.0 multiverse" | tee /etc/apt/sources.list.d/mongodb-org-3.0.list
+	apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 9DA31620334BD75D9DCB49F368818C72E52529D4
+	echo "deb ${PACKAGE_URL} "$(lsb_release -sc)"/mongodb-org/4.0 multiverse" | tee /etc/apt/sources.list.d/mongodb-org-4.0.list
 
 	# Install updates
 	apt-get -y update
@@ -189,8 +193,8 @@ install_mongodb()
 	fi
 	
 	#Install Mongo DB
-	log "Installing MongoDB package $PACKAGE_NAME"
-	apt-get -y install $PACKAGE_NAME
+	log "Installing MongoDB package $PACKAGE_NAME=$PACKAGE_VERSION"
+	apt-get -y install $PACKAGE_NAME=$PACKAGE_VERSION
 	
 	# Stop Mongod as it may be auto-started during the above step (which is not desirable)
 	stop_mongodb
@@ -232,7 +236,7 @@ configure_replicaset()
 		log "Initiating a replica set $REPLICA_SET_NAME with $INSTANCE_COUNT members"
 	
 		# Initiate a replica set
-		mongo master -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host 127.0.0.1 --eval "printjson(rs.initiate())"
+		mongo --authenticationDatabase "admin" -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host 127.0.0.1 --eval "printjson(rs.initiate())"
 		
 		# Add all members except this node as it will be included into the replica set after the above command completes
 		for (( n=0 ; n<($INSTANCE_COUNT-1) ; n++)) 
@@ -240,12 +244,12 @@ configure_replicaset()
 			MEMBER_HOST="${NODE_IP_PREFIX}${n}:${MONGODB_PORT}"
 			
 			log "Adding member $MEMBER_HOST to replica set $REPLICA_SET_NAME" 
-			mongo master -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host 127.0.0.1 --eval "printjson(rs.add('${MEMBER_HOST}'))"
+			mongo --authenticationDatabase "admin" -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host 127.0.0.1 --eval "printjson(rs.add('${MEMBER_HOST}'))"
 		done
 		
 		# Print the current replica set configuration
-		mongo master -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host 127.0.0.1 --eval "printjson(rs.conf())"	
-		mongo master -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host 127.0.0.1 --eval "printjson(rs.status())"	
+		mongo --authenticationDatabase "admin" -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host 127.0.0.1 --eval "printjson(rs.conf())"	
+		mongo --authenticationDatabase "admin" -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host 127.0.0.1 --eval "printjson(rs.status())"	
 	fi
 	
 	# Register an arbiter node with the replica set
@@ -258,7 +262,7 @@ configure_replicaset()
 		CURRENT_NODE_IP=${CURRENT_NODE_IPS[@]}
 
 		log "Adding an arbiter ${HOSTNAME} ($CURRENT_NODE_IP) node to the replica set $REPLICA_SET_NAME"
-		mongo master -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host $PRIMARY_MEMBER_HOST --eval "printjson(rs.addArb('${CURRENT_NODE_IP}'))"
+		mongo --authenticationDatabase "admin" -u $ADMIN_USER_NAME -p $ADMIN_USER_PASSWORD --host $PRIMARY_MEMBER_HOST --eval "printjson(rs.addArb('${CURRENT_NODE_IP}'))"
 	fi
 }
 
@@ -275,10 +279,6 @@ configure_mongodb()
 	chown -R mongodb:mongodb "$MONGODB_DATA/log"
 	chmod 755 "$MONGODB_DATA"
 	
-	mkdir /var/run/mongodb
-	touch /var/run/mongodb/mongod.pid
-	chmod 777 /var/run/mongodb/mongod.pid
-	
 	tee /etc/mongod.conf > /dev/null <<EOF
 systemLog:
     destination: file
@@ -289,6 +289,7 @@ processManagement:
     fork: true
     pidFilePath: "/var/run/mongodb/mongod.pid"
 net:
+    bindIpAll: true
     port: $MONGODB_PORT
 security:
     #keyFile: ""
@@ -304,7 +305,7 @@ EOF
 
 	# Fixing an issue where the mongod will not start after reboot where when /run is tmpfs the /var/run/mongodb directory will be deleted at reboot
 	# After reboot, mongod wouldn't start since the pidFilePath is defined as /var/run/mongodb/mongod.pid in the configuration and path doesn't exist
-	sed -i "s|pre-start script|pre-start script\n  if [ ! -d /var/run/mongodb ]; then\n    mkdir -p /var/run/mongodb \&\& touch /var/run/mongodb/mongod.pid \&\& chmod 777 /var/run/mongodb/mongod.pid\n  fi\n|" /etc/init/mongod.conf
+	sed -i "s|pre-start script|pre-start script\n  if [ ! -d /var/run/mongodb ]; then\n    mkdir -p /var/run/mongodb \&\& touch /var/run/mongodb/mongod.pid \&\& chmod 777 /var/run/mongodb/mongod.pid \&\& chown mongodb:mongodb /var/run/mongodb/mongod.pid\n  fi\n|" /etc/init/mongod.conf
 
 
 }
@@ -337,7 +338,7 @@ configure_db_users()
 {
 	# Create a system administrator
 	log "Creating a system administrator"
-	mongo master --host 127.0.0.1 --eval "db.createUser({user: '${ADMIN_USER_NAME}', pwd: '${ADMIN_USER_PASSWORD}', roles:[{ role: 'userAdminAnyDatabase', db: 'admin' }, { role: 'clusterAdmin', db: 'admin' }, { role: 'readWriteAnyDatabase', db: 'admin' }, { role: 'dbAdminAnyDatabase', db: 'admin' } ]})"
+	mongo admin --host 127.0.0.1 --eval "db.createUser({user: '${ADMIN_USER_NAME}', pwd: '${ADMIN_USER_PASSWORD}', roles:[{ role: 'userAdminAnyDatabase', db: 'admin' }, { role: 'clusterAdmin', db: 'admin' }, { role: 'readWriteAnyDatabase', db: 'admin' }, { role: 'dbAdminAnyDatabase', db: 'admin' } ]})"
 }
 
 # Step 1