+++ postgresql-on-ubuntu/install_postgresql.sh
@@ -10,10 +10,10 @@
 # to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
-# 
+#
 # The above copyright notice and this permission notice shall be included in all
 # copies or substantial portions of the Software.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 # FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -50,7 +50,7 @@ fi
 
 # Get today's date into YYYYMMDD format
 now=$(date +"%Y%m%d")
- 
+
 # Get passed in parameters $1, $2, $3, $4, and others...
 MASTERIP=""
 SUBNETADDRESS=""
@@ -90,17 +90,17 @@ export PGPASSWORD=$REPLICATORPASSWORD
 logger "NOW=$now MASTERIP=$MASTERIP SUBNETADDRESS=$SUBNETADDRESS NODETYPE=$NODETYPE"
 
 install_postgresql_service() {
-	logger "Start installing PostgreSQL..."
+	logger "Start installing PostgreSQL 10..."
 	# Re-synchronize the package index files from their sources. An update should always be performed before an upgrade.
 	apt-get -y update
 
 	# Install PostgreSQL if it is not yet installed
 	if [ $(dpkg-query -W -f='${Status}' postgresql 2>/dev/null | grep -c "ok installed") -eq 0 ];
 	then
-	  apt-get -y install postgresql=9.3* postgresql-contrib=9.3* postgresql-client=9.3*
+	  apt-get -y install postgresql-10 postgresql-contrib-10 postgresql-client-10
 	fi
-	
-	logger "Done installing PostgreSQL..."
+
+	logger "Start installing PostgreSQL 10..."
 }
 
 setup_datadisks() {
@@ -130,7 +130,7 @@ setup_datadisks() {
 
 configure_streaming_replication() {
 	logger "Starting configuring PostgreSQL streaming replication..."
-	
+
 	# Configure the MASTER node
 	if [ "$NODETYPE" == "MASTER" ];
 	then
@@ -156,7 +156,7 @@ configure_streaming_replication() {
 		echo "host replication replicator $SUBNETADDRESS md5" >> pg_hba.conf
 		echo "hostssl replication replicator $SUBNETADDRESS md5" >> pg_hba.conf
 		echo "" >> pg_hba.conf
-			
+
 		logger "Updated pg_hba.conf"
 		echo "Updated pg_hba.conf"
 	fi
@@ -173,12 +173,11 @@ configure_streaming_replication() {
 		echo "wal_level = hot_standby" >> postgresql.conf
 		echo "max_wal_senders = 10" >> postgresql.conf
 		echo "wal_keep_segments = 500" >> postgresql.conf
-		echo "checkpoint_segments = 8" >> postgresql.conf
 		echo "archive_mode = on" >> postgresql.conf
 		echo "archive_command = 'cd .'" >> postgresql.conf
 		echo "hot_standby = on" >> postgresql.conf
 		echo "" >> postgresql.conf
-		
+
 		logger "Updated postgresql.conf"
 		echo "Updated postgresql.conf"
 	fi
@@ -192,17 +191,17 @@ configure_streaming_replication() {
 
 		# Make a binary copy of the database cluster files while making sure the system is put in and out of backup mode automatically
 		logger "Make binary copy of the data directory from master"
-		sudo PGPASSWORD=$PGPASSWORD -u postgres pg_basebackup -h $MASTERIP -D /datadisks/disk1/main -U replicator -x
-		 
+		sudo PGPASSWORD=$PGPASSWORD -u postgres pg_basebackup -h $MASTERIP -D /datadisks/disk1/main -U replicator
+
 		# Create recovery file
 		logger "Create recovery.conf file"
 		cd /var/lib/postgresql/9.3/main/
-		
+
 		sudo -u postgres echo "standby_mode = 'on'" > recovery.conf
 		sudo -u postgres echo "primary_conninfo = 'host=$MASTERIP port=5432 user=replicator password=$PGPASSWORD'" >> recovery.conf
 		sudo -u postgres echo "trigger_file = '/var/lib/postgresql/9.3/main/failover'" >> recovery.conf
 	fi
-	
+
 	logger "Done configuring PostgreSQL streaming replication"
 }