+++ test/template-tests/testcases/deploymentTemplate/Parameters-Property-Must-Exist.test.ps1
@@ -1,0 +1,9 @@
+﻿param(
+[Parameter(Mandatory=$true,Position=0)]
+[PSObject]
+$TemplateObject
+)
+
+if (-not $TemplateObject.psobject.properties.item('parameters')) {
+    throw "Parameters property must exist in the template"
+}