+++ sharepoint-adfs/CHANGELOG.md
@@ -1,4 +1,85 @@
-# Change log for AzureRM template for SharePoint 2016 and 2013 configured with ADFS
+# Change log for Azure template for SharePoint 2019 / 2016 / 2013
+
+## July 2019 update
+
+* Significantly improve reliability of the deployment by mitigating its main source of failures: Add a retry mechanism to resource xRemoteFile when the download fails.
+* Completely configure SharePoint to host and run high-trust provider-hosted add-ins
+* Configure LDAPCP to enable augmentation and remove unused claim types
+* Add the certificate of the domain root authority to the SPTrustedRootAuthority
+* Update apiVersion of all ARM resources to latest version
+* Update some property descriptions in the ARM template
+* Update DSC module SharePointDSC to 3.5
+* Update DSC module xPSDesiredStateConfiguration to 8.8, wich a customization on resource xRemoteFile to deal with random connection errors while downloading LDAPCP
+* Update xActiveDirectory from 2.23 to 3.0
+* Update NetworkingDsc from 6.3 to 7.3
+* Update ActiveDirectoryCSDsc from 3.1 to 3.3
+* Update CertificateDsc from 4.3 to 4.7
+* Update xDnsServer from 1.11 to 1.13
+* Update ComputerManagementDsc from 6.1 to 6.4
+* Update SqlServerDsc from 12.2 to 13.0
+* Update StorageDsc from 4.4 to 4.7
+
+## February 2019 update
+
+* Added SharePoint 2019
+* Added option to enable Hybrid benefit for Windows Server licenses
+* Added option to enable automatic Windows updates
+* SharePoint VMs are now created with no data disk by default, but it can still be added by setting its size
+* Updated SharePointDSC from 2.6 to 3.1, and added unreleased changes of [PR #997](https://github.com/PowerShell/SharePointDsc/pull/997) to fix SPDistributedCacheService error in SharePoint 2019
+* Updated xActiveDirectory from 2.21 to 2.23
+* Updated NetworkingDsc from 6.1 to 6.3
+* Updated ActiveDirectoryCSDsc from 3.0 to 3.1
+* Updated CertificateDsc from 4.2 to 4.3
+* Updated ComputerManagementDsc from 5.2 to 6.1
+* Updated SqlServerDsc from 12.0 to 12.2
+* Updated StorageDsc from 4.0 to 4.4
+* Updated xWebAdministration from 2.2 to 2.4
+
+## November 2018 update
+
+* MySites are now configured as host-named site collections
+* Added a new host-named site collection
+* App catalog is now correctly set
+* Added missing SPN to enable Kerberos authentication to SQL Server
+* Added a wildcard as AdditionalWSFedEndpoint in relying party in ADFS (to support HNSC)
+* Added parameter AdditionalWSFedEndpoint on resource cADFSRelyingPartyTrust of DSC module cADFS
+* Updated name of Azure resources
+* Update API version of all resources
+* Updated resource SPTrustedIdentityTokenIssuer in SharePointDSC 2.5 to support parameter UseWReplyParameter
+* Updated SharePointDSC from 2.2 to 2.6
+* Updated xDnsServer from 1.10 to 1.11
+* Updated xActiveDirectory from 2.18 to 2.21
+* Updated xPSDesiredStateConfiguration from 8.2 to 8.4
+* Updated ActiveDirectoryCSDsc from 2.0 to 3.0
+* Updated ComputerManagementDsc from 5.0 to 5.2
+* Updated SqlServerDsc from 11.2 to 12.0
+* Updated CertificateDsc from 4.0 to 4.2
+* Updated xWebAdministration from 1.20 to 2.2
+* Updated from xNetworking 5.7 to NetworkingDsc 6.1
+
+## June 2018 update
+
+* Removed SharePoint farm account from local administrators group as this is no longer necessary since SharePointDsc 2.2
+* Removed the manual modification in resource xRemoteFile to use TLS 1.2. Fixed that issue properly by setting registry keys in DSC configuration instead
+* SQL Server DatabaseEngine now runs with the SQL service account instead of the machine account
+* Refresh GPOs to ensure CA root cert is present in "cert:\LocalMachine\Root\" before issuing a certificate request
+* Moved all service accounts names from parameters to variables to simplify the template deployment form
+* Updated SharePointDsc to 2.2
+* Updated SqlServerDsc to 11.2
+* Updated ComputerManagementDsc to 5.0
+* Updated xCredSSP to 1.3
+* Updated xNetworking to 5.7
+* Updated xWebAdministration from 1.16 to 1.20
+* Updated xPSDesiredStateConfiguration from 8.0 to 8.2
+* Updated xDnsServer from 1.8 to 1.10
+* Updated from xCertificate 2.8 to CertificateDsc 4.0
+* Updated from xAdcsDeployment 1.1 to ActiveDirectoryCSDsc 2.0
+* Updated xPendingReboot from 0.3 to 0.4
+* Replaced xDisk and cDisk by StorageDsc 4.0
+
+## April 4, 2018 update
+
+* Force protocol TLS 1.2 in Invoke-WebRequest to fix TLS/SSL connection error with GitHub in Windows Server 2012 R2
 
 ## Jaunyary 2018 update